requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.3
django-allauth==0.61.1
django-appconf==1.0.6
django-ckeditor==6.7.1
django-compressor==4.4
django-js-asset==2.2.0
gunicorn==21.2.0
idna==3.6
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
razorpay==1.4.2
rcssmin==1.1.1
requests==2.31.0
requests-oauthlib==2.0.0
rjsmin==1.2.1
setuptools==69.2.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0